test: introduce unit test structure
Move tests for classes to a separate per class file. Use pytest as
testrunner.

Closes: #56
