+++ .travis.yml
@@ -7,5 +7,5 @@ install:
  - pip install -r requirements.txt
  - npm install
 script:
- - python test_model.py
+ - ./test.sh
  - npm run build:deps
