+++ test.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+PYTHONPATH=".:${PYTHONPATH}" pytest test -v -s
